requirements.txt
# Webクローリング・スクレイピング関連
selenium==4.15.2
playwright==1.39.0
beautifulsoup4==4.12.2
requests==2.31.0

# ファイル操作・ユーティリティ
pyyaml==6.0.1
python-dotenv==1.0.0
lxml==4.9.3

# 画像処理
pillow==10.1.0

# LLM連携
openai==1.3.5
langchain==0.0.335

# データ処理
pandas==2.1.3
numpy==1.26.2 